requirements.txt
flask==3.0.0
langchain==1.0.3
langchain-openai==1.0.2
langchain-chroma==1.0.0
chromadb==1.3.4
openai==2.7.1
python-dotenv==1.0.0
python-docx==1.2.0
python-pptx==1.0.2
PyMuPDF==1.26.6
pytesseract==0.3.13
Pillow==10.2.0
langchain-text-splitters==1.0.0

